requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt==5.3.1
psycopg2-binary>=2.9          # for PostgreSQL
django-cors-headers>=4.0
Pillow>=10.0                   # for image uploads (receipts)
python-dotenv>=1.0
pytest
pytest-django
factory-boy